Update content and improve presentation in Introduction-to-Python notebook
The `01-Introduction-to-Python.ipynb` notebook is updated with added markdown cells to provide a more detailed, full-text explanation about Python as a programming language and its characteristics. Existing cells with code and outputs are also modified or removed for an improved and cleaner look of the cell notebook. Unnecessary file `fix_noterbook.py` is deleted.